backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-expert-orchestrator"
version = "0.1.0"
description = "LangGraph Expert Orchestrator - Hierarchical Expertise Router Backend"
authors = [{name = "Guillaume Chevalier", email = "guillaume-chevalier@outlook.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "langgraph==0.5.4",
    "langchain==0.3.27",
    "langchain-openai==0.3.28",
    "pydantic==2.11.7",
    "python-dotenv==1.1.1",
    "openai==1.97.1",
    "httpx==0.28.1",
    # Add test/dev dependencies below if you want them in main install
    # "pytest==7.4.0",
    # "pytest-asyncio==0.21.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
test = [
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "pytest-mock==3.14.1",
    "pytest-xdist==3.5.0",
    "httpx==0.28.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true